requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai
langdetect
python-multipart
python-jose[cryptography]
openai
pillow
aiofiles
psycopg2-binary
python-dotenv